Sidebar
It was created a sidebar with options to see the tasks that will be end
"today" or "later" and see the inbox section, that will be contains the
tasks.

The entities controller was modified to get all the entities from the
DB and by this way, de data will be returned in an array so it can be
possible use the each method and iterate to show the entities names in
the sidebar.