Show specific entity
It was implemented the logic to show an specific entity with her tasks.

entities_controller: Was created a new initialize method to get all the
entities, by this way, all entities can be displayed in the sidebar and
just get the specific tasks depending the view, if the view is the root
view, we get the tasks for the firs tentity, if the view is an specific
view, we get the tasks for that entity.

show view: this file contains the sidebar and the specific entity that
will be showed.

entity file: Was removed the tasks content and it was created a new file
called tasks, and just the contetn will be rendered here.
Note: This file only will display the first entity.

sidebar file: Was adde the route to the link to show an entity.

single_entity file: This file will display the specific entity, and as
same as the entity file, the tasks will be rendered here.

Finally, was created the route to show an specific entity.