Clear entity functionallity completed
It was implemented the logic to clear an entity, this means that all the
tasks related with the entity will be deleted.

The way that this was implemented is the same as the delete
functionality, just a button with the task path with the entity id as
parameter and spefied the method, in this case is delete.

Also, was updated the routes file to add the new route to clear the
entity and the resources to updated and delete a single task.