Locked entity functionality completed.
To make this possible, it was necessary updated the index method in the
entities controller. The index methd gets the first record of entities,
the change was the next:
 - At first, the records was order by date in desc, with the clause
   where statos equals to Active and then get just the first record
   saving them in the "first" variable, by
   this way, in the root path always will be displayed the first entity
   created, if this is deleted, then the next will be displayed.
   If "first" variable has records, we get the tasks for those
   entities, if the "first" variable is nil, that means that does not
   exist any entity or that those entities are locked.

In the entity partial in the shared directory, was added a little form
with a hidden input with the id of the entity taht will be updtaed, and
the same for the single entity partial.

In the routes file was updated the method for the update action from put
to post.